+++ src/test/java/com/github/jlangch/venice/impl/functions/CoreFunctionsTest.java
@@ -1307,6 +1307,27 @@ public class CoreFunctionsTest {
 	}
 	
 	@Test
+	public void test_list_ASTERISK() {
+		final Venice venice = new Venice();
+
+		assertNull(venice.eval("(list* nil)"));
+		
+		assertEquals("(1)", venice.eval("(str (list* [1]))"));
+		assertEquals("(1 2)", venice.eval("(str (list* [1 2]))"));
+		assertEquals("(1 2 3)", venice.eval("(str (list* [1 2 3]))"));
+		
+		assertEquals("(1 9)", venice.eval("(str (list* 1 [9]))"));
+		assertEquals("(1 2 3 8 9)", venice.eval("(str (list* 1 2 3 [8 9]))"));
+		assertEquals("(1 2 3 7 8 9)", venice.eval("(str (list* 1 2 3 [7 8 9]))"));
+		
+		assertEquals("([1 2] 3 7 8 9)", venice.eval("(str (list* [1 2] 3 [7 8 9]))"));
+		
+		assertEquals("(1 2)", venice.eval("(str (list* 1 2 nil))"));
+		assertEquals("(nil 1 2)", venice.eval("(str (list* nil [1 2]))"));
+		assertEquals("(nil 1 2 3)", venice.eval("(str (list* nil 1 [2 3]))"));
+	}
+	
+	@Test
 	public void test_keyword_Q() {
 		final Venice venice = new Venice();